backend/requirements.txt
asgiref==3.11.1
boto3==1.42.76
botocore==1.42.76
Django==5.2.12
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
jmespath==1.1.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
